requirements/base.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
pymongo==3.6.1
flask-mongoengine==0.9.5
mongoengine==0.20.0
marshmallow==3.8.0
bcrypt
Flask-JWT-Extended==3.13.1
PyJWT==1.6.4
